boot2docker: remove ntp, sync with host clock every hour
For filesharing to work properly the VM and host clocks should be in-sync,
even if the host clock has drifted relative to a good external timesource.

This patch removes ntp from the VM and replaces it with an hourly
sync with the VM's RTC which is plumbed into `gmtime`/`localtime`
in `xhyve`.

Fixes [docker/pinata#229]

Signed-off-by: David Scott <dave.scott@unikernel.com>