Run qemu-system-x86_64 with more memory.
With the default 128M of memory, the ramfs (which currently has a decompressed
size of ~82M) can't be completely written with the kernel and the compressed ramfs
also in memory, resulting in missing chunks of filesystem (notably /etc and /sbin for
me) on the emulated machine.  This makes for a dramatically broken moby.  Asking for 256M at least
gets us to some nice whale ASCII art and a console we can interact with.