Use IP rather than socket9p for the docker socket
Before this patch, docker would only listen on 0.0.0.0:2375 if running
on Hyper-V. This patch makes it listen there on all hypervisors.

In the case of xhyve, it will listen on the host internal vmnet network
or the virtual slirp network. We should aim to replace this with a socket-
based transport instead.

Signed-off-by: David Scott <dave.scott@docker.com>