Use $ntpsrv from DHCP to configure chrony
The existing `ntp=gateway` cmdline support overrides this, but is largely
redundant now.

A DHCP derived server will not be given the "trust" directive in the
configuration file, since it is harder to argue that we know we should trust a
server which arrives this way since it depends on the context in which Moby is
running. It's not clear to me that it makes much difference anyway.

This only uses the first server in $ntpsrv, since Piñata will only provide one
and supporting more makes the sed quite a bit more complex.

If `ntp=gateway` is given then we now use $router instead of parsing IP output.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>