kernel: Replace vsock email series with upstream patches
The vsock patches were merged into mainline in v4.8-rc1, replace the patches
based on the email posting with cherry-picks. The patches appear identical so
the only change here is to the annotation of the origin which now references an
upstream commit instead of a message-id.

Add one new upstream patch b226acab2f6a "VSOCK: Use kvfree()"

Signed-off-by: Ian Campbell <ian.campbell@docker.com>