Enable local service
Files in `/etc/local.d/` will be executed:

If a file in this directory is executable and it has a .start extension,
it will be run when the local service is started. If a file is
executable and it has a .stop extension, it will be run when the local
service is stopped.

Signed-off-by: Justin Cormack <justin.cormack@docker.com>