efi: add 'noautodetect' to kernel commandline
This stops device auto-detect in 'hwdrivers'. On Docker
for Windows driver auto-detect should not be necessary as all
drivers are compiled in.

Further, on slow machines (and in CI) th 'hwdrivers' seems to have
race as it does a 'find' on '/sys/devices' and then performs actions.
On slow machines (and in CI) it appears that VMBus devices get
removed (and possibly re-added) causing the script to error.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>