containerd: Don't use standard NATS port
Newer versions of containerd provide a pub/sub based service
to things like engine. By default this uses the standrd NATS
port 4222 which interferes with users wanting to run NATS in
a container.

This commit moves the NATS port used by containerd to a different
port. No other changes should be required as we run containerd
in standalone mode.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>