switch CONFIG_PAGE_POISON_NO_SANITY=n
This has a non-trivial performance impact, since now we're verifying the
page poison values on each page allocation in addition to poisoning the
pages on free. However, this may also prevent some classes of attacks (e.g.
the attacker has a use after free of something and can write to freed pages
somehow).

Signed-off-by: Tycho Andersen <tycho@docker.com>