pkg: Consolidate the rngd package under ./pkg/rngd
The build of the rngd package was split between ./base/rng-tools,
which was used to build the binaries, and ./pkg/rngd which packaged
the binary into a daemon container.

Since the rng-tools binaries are only used by the rngd package,
there is no need to push them to hub as a intermediary stage and
it makes sense to consolidate the building and packaging in ./pkg.

The Makefile is also better documented to hopefully serve as a
template/example for future packages.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>