makefile: Allow a bin/moby to be provided
This is needed for CI to be able to run `moby run $platform` with a
known good version of moby. For example, the one generated when building
the `master`.

This protects CI from any nefarious changes to the `moby` tool and
allows `moby run` to be used safely to test PRs.

Signed-off-by: Dave Tucker <dt@docker.com>