cmd: Enable virtio sockets for HyperKit
HyperKit optionally exposes virtio sockets via a named pipes
in a directory. This patch passes a state directory and sets
the VSock flag to enable this feature.

Note, this will create a directory called 'prefix'-hyperkit
on every execution of 'linuxkit run hyperkit'.

While at it, also fix the .gitignore settings for kernels.

Signed-off-by: Rolf Neugebauer <rolf.neugebauer@docker.com>